Added Exception Handling

+++ serverside/src/main/java/com/onlinebanking/serverside/service/AdminService.java
@@ -3,7 +3,10 @@ package com.onlinebanking.serverside.service;
 import java.util.List;
 
 import com.onlinebanking.serverside.dao.CustomerRepository;
+import com.onlinebanking.serverside.exceptions.AccountNotFoundException;
+import com.onlinebanking.serverside.exceptions.CustomerNotFoundException;
 import com.onlinebanking.serverside.model.Customer;
+import org.checkerframework.checker.units.qual.C;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -54,14 +57,19 @@ public class AdminService {
         return count == 1 ? true : false;
     }
 
-    public boolean deleteAccount(long accNo) {
+    public boolean deleteAccount(long accNo) throws AccountNotFoundException {
         Account account = accRepository.findByAccNo(accNo);
+        if (account == null)
+            throw new AccountNotFoundException("No account exists with the given account number!");
         boolean res = transactionService.deleteTransactionByAccNo(account);
         accRepository.deleteByAccNo(accNo);
         return (accRepository.findByAccNo(accNo) == null);
     }
 
-    public boolean editCustomer(Customer customer) {
+    public boolean editCustomer(Customer customer) throws CustomerNotFoundException {
+        Customer response = customerRepository.findByUserId(customer.getUserId());
+        if (response == null)
+            throw new CustomerNotFoundException("No customer exists with the provided user Id!");
         Customer editedCustomer = customerRepository.save(customer);
         return true;
     }
